+++ tomviz/PythonUtilities.h
@@ -107,7 +107,7 @@ public:
     Dict(const Dict& other);
     Dict(const Object& obj);
     Dict& operator=(const Object& other);
-    ssize_t size() const;
+    long long size() const;
     QStringList keys() const;
     bool delItem(const QString& key);
     Object operator[](const QString& key);
